+++ client/src/game/ScoresZone.java
@@ -1,32 +1,31 @@
 package game;
 
 import java.awt.Color;
-import java.awt.Dimension;
 
-import javax.swing.DefaultListModel;
 import javax.swing.JList;
 import javax.swing.JScrollPane;
 import javax.swing.border.EmptyBorder;
 
 /**
  * Classe définissant une zone d'affichage pour les scores.
- * On part ici du principe qu'il s'agit d'un objet de type JList.
- * 
+ * Il s'agit ici d'un conteneur avec ascenseur vertical.
+ * Utile si le jeu accueille un grand nombre de joueurs (car les scores sortent alors de l'écran)
+ *  
  * @author Baptiste Vannesson
  */
 public class ScoresZone extends JScrollPane {
     
     /** Constante déterminant la marge interne supérieure */
-    public static final int PADDING_TOP = 0;
+    public static final byte PADDING_TOP = 0;
     
     /** Constante déterminant la marge interne gauche */
-    public static final int PADDING_LEFT = 0;
+    public static final byte PADDING_LEFT = 0;
     
     /** Constante déterminant la marge interne inférieure */
-    public static final int PADDING_BOTTOM = 0;
+    public static final byte PADDING_BOTTOM = 0;
     
     /** Constante déterminant la marge interne droite */
-    public static final int PADDING_RIGHT = 3;
+    public static final byte PADDING_RIGHT = 3;
     
     /**
      * Constructeur de la zone d'affichage des scores.
